+++ test/jest/__mock__/reactIntl.mock.js
@@ -1,6 +1,7 @@
 jest.mock('react-intl', () => {
   const intl = {
     formatMessage: ({ id }) => id,
+    locale: 'en'
   };
 
   return {